refactor: share observable seed processing
